requirements.txt
aiofiles==23.2.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.3.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
Django==4.2.11
django-apscheduler==0.6.2
django-compat==1.0.15
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kombu==5.3.7
openai==1.37.0
packaging==24.0
pillow==10.4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
pytz==2024.1
redis==5.0.4
requests==2.31.0
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
xmltodict==0.13.0
zope.event==5.0
zope.interface==6.3